added error handling and ways to call refresh token endpoints in response to 401 or 403 errors
